feat: support tilde in url
From Roberta Calegari <roberta.calegari@unibo.it>

May I kindly ask you also if it is possible to insert the  ~ in the url of authors command?
We have a problem with the authors urls in the new format: the tilde ~ symbol that is interpreted in a very strange way, generating a bad url, and it's not clear to me how to solve the issue.

\author[4]{Amal {El Fallah-Seghrouchni}}[%
orcid=0000-0002-8390-8780,
email=amal.elfallah@lip6.fr,
url={https://webia.lip6.fr/~elfallah/},
]

Produces the url https://webia.lip6.fr/\protect\unhbox\voidb@x\protect\penalty\@M\{}elfallah/

I tried with url=https://webia.lip6.fr/\%7Eelfallah/, but it prints the %7E symbols, I tried with \texttildelow,  \~, ~{}, and many other options, but some problems are still there.